+++ src/root/utils/utils.js
@@ -3,6 +3,7 @@ import _groupBy from 'lodash.groupby';
 import _toNumber from 'lodash.tonumber';
 import _find from 'lodash.find';
 import _filter from 'lodash.filter';
+import * as EmailValidator from 'email-validator';
 import { DateTime } from 'luxon';
 import { isNotDefined } from '@togglecorp/fujs';
 
@@ -153,25 +154,18 @@ export function mostRecentReport (reports) {
 }
 
 export function isValidEmail (email) {
-  // https://stackoverflow.com/a/1373724
-  // The official standard is known as RFC 2822. It describes the syntax that valid email addresses must adhere to. You can (but you shouldn't — read on) implement it with this regular expression:
-
-  // (?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])
-
-  // (...) We get a more practical implementation of RFC 2822 if we omit the syntax using double quotes and square brackets. It will still match 99.99% of all email addresses in actual use today.
-
-  // [a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?
-
-  // A further change you could make is to allow any two-letter country code top level domain, and only specific generic top level domains. This regex filters dummy email addresses like asdf@adsf.adsf. You will need to update it as new top-level domains are added.
-
-  // [a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+(?:[A-Z]{2}|com|org|net|gov|mil|biz|info|mobi|name|aero|jobs|museum)\b
-
-  // So even when following official standards, there are still trade-offs to be made. Don't blindly copy regular expressions from online libraries or discussion forums. Always test them on your own data and with your own applications.
-  return /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/.test(email);
+  return EmailValidator.validate(email);
 }
 
 export function isWhitelistedEmail (email, whitelistedDomains) {
-  return isValidEmail(email) && whitelistedDomains.find(o => email.indexOf(`@${o}`) !== -1);
+  if (!isValidEmail(email)) {
+    return false;
+  }
+
+  // Looking for an EXACT match in the domain whitelist
+  // (it finds even if UPPERCASE letters were used)
+  const userMailDomain = email.substring(email.lastIndexOf("@") +1);
+  return whitelistedDomains.find(dom => dom === userMailDomain);
 }
 
 export function finishedFetch (curr, next, prop) {