+++ src/root/views/Covid19ThreeWSankey/Filters.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import Faram from '@togglecorp/faram';
+import _cs from 'classnames';
 
 import SelectInput from '../../components/form-elements/select-input';
 import { countryNameMapByIso } from '../../utils/field-report-constants';
+
+import styles from './styles.module.scss';
 
 const filterSchema = {
   fields: {
@@ -30,6 +33,7 @@ function NSActivitiesFilters (p) {
     value,
     onChange,
     data = emptyData,
+    className,
   } = p;
 
   const [
@@ -51,7 +55,7 @@ function NSActivitiesFilters (p) {
 
   return (
     <Faram
-      className='regional-ns-activity-filter'
+      className={_cs(className, styles.filters)}
       schema={filterSchema}
       value={value}
       onChange={onChange}
@@ -61,7 +65,7 @@ function NSActivitiesFilters (p) {
         label='Supporting NS'
         placeholder='Select Who (Reporting NS)'
         options={supportingNSOptions}
-        className='select-input'
+        className={styles.input}
         multi
       />
       <SelectInput
@@ -69,7 +73,7 @@ function NSActivitiesFilters (p) {
         label='Activity'
         placeholder='Select What (Sector)'
         options={activityOptions}
-        className='select-input'
+        className={styles.input}
         multi
       />
       <SelectInput
@@ -77,7 +81,7 @@ function NSActivitiesFilters (p) {
         label='Receiving NS'
         placeholder='Select Where (Country)'
         options={receivingNSOptions}
-        className='select-input'
+        className={styles.input}
         multi
       />
     </Faram>