+++ src/root/views/field-report-form/index.js
@@ -450,7 +450,7 @@ class FieldReportForm extends React.Component {
               placeholder='Select a disaster type'
               name='disaster-type'
               id='disaster-type'
-              disabled={ this.state.data.isCovidReport === 'true' }
+              disabled={ this.state.data.isCovidReport === 'true' || !this.state.data.isCovidReport }
               options={formData.disasterType}
               value={this.state.data.disasterType}
               onChange={({value}) => this.onFieldChange('disasterType', value)}
@@ -718,10 +718,12 @@ class FieldReportForm extends React.Component {
               });
               values.options = sectionValues;
 
+              const description = this.state.data.isCovidReport === 'true' ? section.desc[status + '-COV'] : section.desc[status];
+
               return (
                 <ActionsCheckboxes
                   label={section.label[status]}
-                  description={section.desc[status]}
+                  description={description}
                   placeholder={section.placeholder[status]}
                   name={section.name}
                   key={section.key}